refactor: old prior fit implementation moved to tests.
